server/pyproject.toml
[project]
name = "mcp-bot"
version = "0.1.0"
description = "A bot for simplify the abstraction from technical manuals"
authors = [
    {name = "Samuel Jauregui",email = "samjaurz_@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
package-mode = false
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "gpt4all (>=2.8.2,<3.0.0)",
    "llama-cpp-python (==0.1.78)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "anthropic (>=0.68.0,<0.69.0)",
    "black (>=25.1.0,<26.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
